Restructuring header files
