+++ headers/Archer.h
@@ -2,7 +2,7 @@
 
 #ifndef ARCHER_H
 #define ARCHER_H
-#include "../Role.h"
+#include "Role.h"
 #include <random>
 class Archer : public Role {
 public:
