+++ headers/teamCreator.h
@@ -6,25 +6,25 @@
 #include <string>
 #include <unordered_map>
 #include <random>
-#include "../../src/Role.h"
-#include "../../src/Team/Team.h"
-#include "../../src/magic_roles/Alchemist.h"
-#include "../../src/magic_roles/Druid.h"
-#include "../../src/magic_roles/Mage.h"
-#include "../../src/magic_roles/Necromancer.h"
-#include "../../src/magic_roles/Shaman.h"
-#include "../../src/magic_roles/Sorcerer.h"
-#include "../../src/Gameplay/GameReloader.h"
-#include "../../src/Role.h"
-#include "../../src/agility_roles/Archer.h"
-#include "../../src/agility_roles/Assassin.h"
-#include "../../src/agility_roles/Doctor.h"
-#include "../../src/agility_roles/Ecnchantress.h"
-#include "../../src/agility_roles/Monk.h"
-#include "../../src/agility_roles/Thief.h"
-#include "../../src/ancherman_roles/Gladiator.h"
-#include "../../src/ancherman_roles/Knight.h"
-#include "../../src/ancherman_roles/Warrior.h"
+#include "Role.h"
+#include "Team.h"
+#include "Alchemist.h"
+#include "Druid.h"
+#include "Mage.h"
+#include "Necromancer.h"
+#include "Shaman.h"
+#include "Sorcerer.h"
+#include "GameReloader.h"
+#include "Role.h"
+#include "Archer.h"
+#include "Assassin.h"
+#include "Doctor.h"
+#include "Ecnchantress.h"
+#include "Monk.h"
+#include "Thief.h"
+#include "Gladiator.h"
+#include "Knight.h"
+#include "Warrior.h"
 class TeamCreator {
 private:
     const std::unordered_map<std::string, std::vector<int>> roleBaseStats = {
