+++ src/Gameplay/GameReloader.cpp
@@ -1,24 +1,24 @@
-#include "GameReloader.h"
-#include "../Role.h"
+#include "../../headers/GameReloader.h"
+#include "../../headers/Role.h"
 #include <fstream>
 #include <vector>
 #include <memory>
 #include <stdexcept>
-#include "../agility_roles/Archer.h"
-#include "../agility_roles/Assassin.h"
-#include "../agility_roles/Doctor.h"
-#include "../agility_roles/Ecnchantress.h"
-#include "../agility_roles/Monk.h"
-#include "../agility_roles/Thief.h"
-#include "../ancherman_roles/Gladiator.h"
-#include "../ancherman_roles/Knight.h"
-#include "../ancherman_roles/Warrior.h"
-#include "../magic_roles/Alchemist.h"
-#include "../magic_roles/Druid.h"
-#include "../magic_roles/Mage.h"
-#include "../magic_roles/Necromancer.h"
-#include "../magic_roles/Shaman.h"
-#include "../magic_roles/Sorcerer.h"
+#include "../../headers/Archer.h"
+#include "../../headers/Assassin.h"
+#include "../../headers/Doctor.h"
+#include "../../headers/Ecnchantress.h"
+#include "../../headers/Monk.h"
+#include "../../headers/Thief.h"
+#include "../../headers/Gladiator.h"
+#include "../../headers/Knight.h"
+#include "../../headers/Warrior.h"
+#include "../../headers/Alchemist.h"
+#include "../../headers/Druid.h"
+#include "../../headers/Mage.h"
+#include "../../headers/Necromancer.h"
+#include "../../headers/Shaman.h"
+#include "../../headers/Sorcerer.h"
 #include <sstream>
 #include <string>
 std::unique_ptr<Role> createGameRole(const std::string& line) {
