--- src/Gameplay/Gameplay.cpp
@@ -1,5 +1,0 @@
-//
-// Created by ruslan on 3/6/25.
-//
-
-#include "Gameplay.h"
