--- src/Gameplay/Gameplay.h
@@ -1,12 +1,0 @@
-
-#ifndef GAMEPLAY_H
-#define GAMEPLAY_H
-#include "../Role.h"
-
-class Gameplay : public Role {};
-
-void gameManagemnet();
-void gameLive();
-
-
-#endif //GAMEPLAY_H
