--- src/Gameplay/out.txt
@@ -1,2 +1,0 @@
-5 5 5 250 20 0 1 f
-5 5 5 200 10 0 1 t
