+++ src/Team/Team.cpp
@@ -2,7 +2,7 @@
 // Created by ruslan on 3/6/25.
 //
 
-#include "Team.h"
+#include "../../headers/Team.h"
 #include <iostream>
